pyproject.toml
[tool.poetry]
name = "crawler_server"
version = "0.1.0"
description = "Server for Mwmbl's web crawler"
authors = ["Daoud Clarke <daoud.clarke@gmail.com>"]
license = "AGPL"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.71.0"
uvicorn = "^0.16.0"
boto3 = "^1.20.37"
requests = "^2.27.1"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
